Reorganize example source structure
