+++ example/src/resources/slider-left.png
