+++ example/src/resources/slider-right.png
