+++ example/src/resources/slider.png
